% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{DGP_estimation}
\alias{DGP_estimation}
\title{DGP estimation
Notes on how to generalize:
\itemize{
\item assume time ordering of columns
\item identify baseline covariates
\item identify tv covariates
\item generate lagged data
\item fit models to all history (use sl3)
}}
\usage{
DGP_estimation(data)
}
\description{
DGP estimation
Notes on how to generalize:
\itemize{
\item assume time ordering of columns
\item identify baseline covariates
\item identify tv covariates
\item generate lagged data
\item fit models to all history (use sl3)
}
}
