% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{fit_column}
\alias{fit_column}
\title{Fit a single column based on all previous columns, excluding id}
\usage{
fit_column(pred_data, col, learner = NULL, counting = FALSE, verbose = TRUE)
}
\description{
Fit a single column based on all previous columns, excluding id
}
