% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pascLtmle.R
\name{pascLtmle}
\alias{pascLtmle}
\title{PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)}
\description{
PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)

PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)
}
\section{Super class}{
\code{tmle3sim::t3s_Estimator} -> \code{pascLtmle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pascLtmle-new}{\code{pascLtmle$new()}}
\item \href{#method-pascLtmle-estimate}{\code{pascLtmle$estimate()}}
\item \href{#method-pascLtmle-clone}{\code{pascLtmle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Estimator" data-id="print"><a href='../../tmle3sim/html/t3s_Estimator.html#method-t3s_Estimator-print'><code>tmle3sim::t3s_Estimator$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascLtmle-new"></a>}}
\if{latex}{\out{\hypertarget{method-pascLtmle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascLtmle$new(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascLtmle-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-pascLtmle-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascLtmle$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascLtmle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pascLtmle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascLtmle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
