% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PascSim.R
\name{PascSim}
\alias{PascSim}
\title{PASC TRUE DGD}
\description{
PASC TRUE DGD

PASC TRUE DGD
}
\section{Super class}{
\code{tmle3sim::t3s_Simulation} -> \code{PascSim}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PascSim-new}{\code{PascSim$new()}}
\item \href{#method-PascSim-sample}{\code{PascSim$sample()}}
\item \href{#method-PascSim-clone}{\code{PascSim$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Simulation" data-id="print"><a href='../../tmle3sim/html/t3s_Simulation.html#method-t3s_Simulation-print'><code>tmle3sim::t3s_Simulation$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Simulation" data-id="run"><a href='../../tmle3sim/html/t3s_Simulation.html#method-t3s_Simulation-run'><code>tmle3sim::t3s_Simulation$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Simulation" data-id="run_step"><a href='../../tmle3sim/html/t3s_Simulation.html#method-t3s_Simulation-run_step'><code>tmle3sim::t3s_Simulation$run_step()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PascSim-new"></a>}}
\if{latex}{\out{\hypertarget{method-PascSim-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PascSim$new(params = list(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PascSim-sample"></a>}}
\if{latex}{\out{\hypertarget{method-PascSim-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PascSim$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PascSim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PascSim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PascSim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
