% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pascStremr.R
\name{pascStemr}
\alias{pascStemr}
\title{PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)}
\description{
PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)

PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)
}
\section{Super class}{
\code{tmle3sim::t3s_Estimator} -> \code{pascStremr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pascStremr-new}{\code{pascStemr$new()}}
\item \href{#method-pascStremr-estimate}{\code{pascStemr$estimate()}}
\item \href{#method-pascStremr-clone}{\code{pascStemr$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Estimator" data-id="print"><a href='../../tmle3sim/html/t3s_Estimator.html#method-t3s_Estimator-print'><code>tmle3sim::t3s_Estimator$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascStremr-new"></a>}}
\if{latex}{\out{\hypertarget{method-pascStremr-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$new(params = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascStremr-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-pascStremr-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pascStremr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pascStremr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
