% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_Reporter.R
\name{ps_Reporter}
\alias{ps_Reporter}
\title{PASC Sim Reporting}
\description{
PASC Sim Reporting

PASC Sim Reporting
}
\section{Super class}{
\code{tmle3sim::t3s_Reporter} -> \code{ps_Reporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ps_Reporter-new}{\code{ps_Reporter$new()}}
\item \href{#method-ps_Reporter-report}{\code{ps_Reporter$report()}}
\item \href{#method-ps_Reporter-make_final}{\code{ps_Reporter$make_final()}}
\item \href{#method-ps_Reporter-clone}{\code{ps_Reporter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Reporter" data-id="print"><a href='../../tmle3sim/html/t3s_Reporter.html#method-t3s_Reporter-print'><code>tmle3sim::t3s_Reporter$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Reporter" data-id="save"><a href='../../tmle3sim/html/t3s_Reporter.html#method-t3s_Reporter-save'><code>tmle3sim::t3s_Reporter$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_Reporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ps_Reporter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$new(params = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_Reporter-report"></a>}}
\if{latex}{\out{\hypertarget{method-ps_Reporter-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_Reporter-make_final"></a>}}
\if{latex}{\out{\hypertarget{method-ps_Reporter-make_final}{}}}
\subsection{Method \code{make_final()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$make_final()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_Reporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ps_Reporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
