% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthetic_sim_functions.R
\name{repeated_data}
\alias{repeated_data}
\title{Add history data from previous two timepoints}
\usage{
repeated_data(long_data, pseudo)
}
\description{
Add history data from previous two timepoints
}
