% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticData.R
\name{SyntheticData}
\alias{SyntheticData}
\title{PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)}
\description{
PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)

PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)
}
\section{Super class}{
\code{tmle3sim::t3s_Estimator} -> \code{SyntheticData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SyntheticData-new}{\code{SyntheticData$new()}}
\item \href{#method-SyntheticData-estimate}{\code{SyntheticData$estimate()}}
\item \href{#method-SyntheticData-clone}{\code{SyntheticData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Estimator" data-id="print"><a href='../../tmle3sim/html/t3s_Estimator.html#method-t3s_Estimator-print'><code>tmle3sim::t3s_Estimator$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SyntheticData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SyntheticData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyntheticData$new(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SyntheticData-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SyntheticData-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyntheticData$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SyntheticData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SyntheticData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SyntheticData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
