% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_baseline}
\alias{generate_baseline}
\alias{generate_tv_init}
\alias{generate_tv_single}
\alias{generate_tv_all}
\alias{generate_data}
\title{DGD Functions}
\usage{
generate_baseline(n)

generate_tv_init(baseline_covariates)

generate_tv_single(
  baseline_covariates,
  current_tv,
  new_time,
  regime = NA,
  limit_covid = TRUE,
  effect_size
)

generate_tv_all(
  baseline_covariates,
  regime,
  time_increment,
  end_time,
  limit_covid,
  effect_size
)

generate_data(
  n = 1000,
  full = TRUE,
  include_p = TRUE,
  regime = NA,
  time_increment = 6,
  end_time = 20 * 30,
  obs_scale = 30,
  limit_covid = TRUE,
  effect_size = 1/30,
  ...
)
}
\description{
DGD Functions
}
