% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthetic_sim_functions.R
\name{DGP_estimation}
\alias{DGP_estimation}
\title{DGP estimation
Notes on how to generalize:
\itemize{
\item identify baseline covariates
\item 
}}
\usage{
DGP_estimation(sim_result)
}
\description{
DGP estimation
Notes on how to generalize:
\itemize{
\item identify baseline covariates
\item 
}
}
