% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pascStremr.R
\name{pascStemr}
\alias{pascStemr}
\title{PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)}
\description{
PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)

PASC Sim Synthetic Data Estimation (this should probably yield a new sim we can sample from and apply the same estimation)
}
\section{Super class}{
\code{\link[tmle3sim:t3s_Estimator]{tmle3sim::t3s_Estimator}} -> \code{pascStremr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pascStemr$new()}}
\item \href{#method-estimate}{\code{pascStemr$estimate()}}
\item \href{#method-clone}{\code{pascStemr$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Estimator" data-id="print">}\href{../../tmle3sim/html/t3s_Estimator.html#method-print}{\code{tmle3sim::t3s_Estimator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$new(params = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pascStemr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
