% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_Reporter.R
\name{ps_Reporter}
\alias{ps_Reporter}
\title{PASC Sim Reporting}
\description{
PASC Sim Reporting

PASC Sim Reporting
}
\section{Super class}{
\code{\link[tmle3sim:t3s_Reporter]{tmle3sim::t3s_Reporter}} -> \code{ps_Reporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ps_Reporter$new()}}
\item \href{#method-report}{\code{ps_Reporter$report()}}
\item \href{#method-make_final}{\code{ps_Reporter$make_final()}}
\item \href{#method-clone}{\code{ps_Reporter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Reporter" data-id="print">}\href{../../tmle3sim/html/t3s_Reporter.html#method-print}{\code{tmle3sim::t3s_Reporter$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3sim" data-topic="t3s_Reporter" data-id="save">}\href{../../tmle3sim/html/t3s_Reporter.html#method-save}{\code{tmle3sim::t3s_Reporter$save()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$new(params = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report"></a>}}
\if{latex}{\out{\hypertarget{method-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_final"></a>}}
\if{latex}{\out{\hypertarget{method-make_final}{}}}
\subsection{Method \code{make_final()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$make_final()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ps_Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
